Remove setup in separate KlaviyoAction, consolidate to call in auth path
